refactor: remove testConfig from lib/utils
