+++ lib/utils/index.ts
@@ -28,12 +28,3 @@ export const parseNamespace = (namespace: unknown): string => {
 
     return 'unknown';
 };
-
-/**
- * Uses for testing.
- */
-export const testConfig = {
-    host: process.env.REDIS_HOST ?? '127.0.0.1',
-    port: Number.parseInt(process.env.REDIS_PORT ?? '6379', 10),
-    password: process.env.REDIS_PASSWORD
-};
