+++ test/redis/app/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TerminusModule } from '@nestjs/terminus';
 import { RedisModule } from '../../../lib';
-import { testConfig } from '../../../lib/utils';
+import { testConfig } from '../../../jest-env';
 import { HealthController } from './controllers/health.controller';
 import { InjectController } from './controllers/inject.controller';
 import { ServiceController } from './controllers/service.controller';
@@ -12,7 +12,9 @@ export const CLIENT_0 = 'client0';
     imports: [
         TerminusModule,
         RedisModule.forRoot({
-            defaultOptions: testConfig,
+            defaultOptions: {
+                ...testConfig.master
+            },
             config: [{ namespace: 'client0', db: 0 }, { db: 1 }]
         })
     ],
